--- docs/model/univnet/generator.md
@@ -1,1 +1,0 @@
-::: model.univnet.generator
