--- docs/model/univnet/log_stft_magnitude_loss.md
@@ -1,1 +1,0 @@
-::: model.univnet.log_stft_magnitude_loss
