--- docs/model/univnet/multi_resolution_stft_loss.md
@@ -1,1 +1,0 @@
-::: model.univnet.multi_resolution_stft_loss
