--- docs/model/univnet/spectral_convergence_loss.md
@@ -1,1 +1,0 @@
-::: model.univnet.spectral_convergence_loss
