--- docs/model/univnet/stft.md
@@ -1,1 +1,0 @@
-::: model.univnet.stft
