--- docs/model/univnet/stft_loss.md
@@ -1,1 +1,0 @@
-::: model.univnet.stft_loss
