--- docs/model/univnet/traced_generator.md
@@ -1,1 +1,0 @@
-::: model.univnet.traced_generator
