+++ docs/model/univnet/univnet.md
@@ -1,0 +1,1 @@
+::: model.univnet.univnet
