+++ docs/training/loss/multi_resolution_stft_loss.md
@@ -1,0 +1,1 @@
+::: training.loss.multi_resolution_stft_loss
