+++ docs/training/loss/stft.md
@@ -1,0 +1,1 @@
+::: training.loss.stft
