+++ docs/training/loss/stft_loss.md
@@ -1,0 +1,1 @@
+::: training.loss.stft_loss
