+++ training/loss/__init__.py
@@ -1,4 +1,8 @@
 from .bin_loss import BinLoss
 from .fast_speech_2_loss_gen import FastSpeech2LossGen
 from .forward_sum_loss import ForwardSumLoss
+from .log_stft_magnitude_loss import LogSTFTMagnitudeLoss
 from .losses_criterion import LossesCriterion
+from .multi_resolution_stft_loss import MultiResolutionSTFTLoss
+from .spectral_convergence_loss import SpectralConvergengeLoss
+from .stft_loss import STFTLoss
