+++ training/loss/tests/test_spectral_convergence_loss.py
@@ -2,25 +2,26 @@ import unittest
 
 import torch
 
-from model.univnet.log_stft_magnitude_loss import LogSTFTMagnitudeLoss
+from training.loss import SpectralConvergengeLoss
 
 
-class TestLogSTFTMagnitudeLoss(unittest.TestCase):
-    def test_log_stft_magnitude_loss(self):
-        # Test the log STFT magnitude loss function with random input tensors
-        loss_fn = LogSTFTMagnitudeLoss()
+class TestSpectralConvergengeLoss(unittest.TestCase):
+    def test_spectral_convergence_loss(self):
+        # Test the spectral convergence loss function with random input tensors
+        loss_fn = SpectralConvergengeLoss()
 
         x_mag = torch.randn(4, 100, 513)
-        y_mag = torch.randn(4, 100, 513)
+        y_mag = torch.randn(4, 100, 513) * 0.1
 
         loss = loss_fn(x_mag, y_mag)
 
         self.assertIsInstance(loss, torch.Tensor)
         self.assertEqual(loss.shape, torch.Size([]))
+        self.assertGreater(loss, 0.0)
 
-    def test_log_stft_magnitude_loss_nonzero(self):
-        # Test the log STFT magnitude loss function with non-zero loss
-        loss_fn = LogSTFTMagnitudeLoss()
+    def test_spectral_convergence_small_vectors(self):
+        # Test the spectral convergence loss function with non-zero loss
+        loss_fn = SpectralConvergengeLoss()
 
         x_mag = torch.tensor([1, 4, 9, 64], dtype=torch.float32)
         y_mag = torch.tensor([1, 8, 16, 256], dtype=torch.float32)
@@ -30,7 +31,7 @@ class TestLogSTFTMagnitudeLoss(unittest.TestCase):
         self.assertIsInstance(loss, torch.Tensor)
         self.assertEqual(loss.shape, torch.Size([]))
 
-        expected = torch.tensor(0.6637)
+        expected = torch.tensor(0.7488)
         self.assertTrue(torch.allclose(loss, expected, rtol=1e-4, atol=1e-4))
